requirements.txt
annotated-types==0.7.0
anyio==4.4.0
catboost==1.2.5
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
fonttools==4.53.1
graphviz==0.20.3
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
numpy==1.26.4
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pillow==10.4.0
plotly==5.23.0
psycopg2==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
tenacity==8.5.0
threadpoolctl==3.5.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.3
watchfiles==0.22.0
websockets==12.0
xgboost==2.1.0
